UPSTREAM: arm64: vdso32: Don't use KBUILD_CPPFLAGS unconditionally
KBUILD_CPPFLAGS is defined differently depending on whether the main
compiler is clang or not. This means that it is not possible to build
the compat vDSO with GCC if the rest of the kernel is built with clang.

Define VDSO_CPPFLAGS directly to break this dependency and allow a clang
kernel to build a compat vDSO with GCC:

  $ make ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- \
    CROSS_COMPILE_COMPAT=arm-linux-gnueabihf- CC=clang \
    COMPATCC=arm-linux-gnueabihf-gcc

Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit c71e88c437962c1ec43d4d23a0ebf4c9cf9bee0d)
Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 154668398
Change-Id: I43e54c145831fde92e6883004e9addcb6cbe3ef7
(cherry-picked from commit 0edaa642358b9a72f9a3bca30c0b66b7d396dcfd)
Signed-off-by: Albert I <kras@raphielgang.org>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>